Implement method to shut down a program when CPU too hot
The KillOCTestApp method has been added to kill an app if the
any of the CPU's core temps reach a certain threshold.  For now,
the limit has been set to 20 degrees Celcius, and the app we
are looking to shutdown is Google Chrome.  The objective is
do eventually have a user specfiy which app they want killed
before running the temp checks.

GetSystemInfo was renamed to IsCPUTooHot, and now returns a bool.
This will help in determining when we need to run the KillOCTestApp
method.