Display reason for termination to user.  Remove unneeded file.
There are two ways a program can end while runninging FireDoor; a
user can terminate the program they are actively testing or
FireDoor can shutdown the program when the max temprature
threshold is reached.  The app was set to detect either of these
cases, and return a result to the user indicated what shut down
the test program.  If FireDoor shut the program down, it will
include which core triggered the shutdown and what temprature it
was.

apps.json was also removed, as we are allowing users to define the
path to their test program upon startup.