Check to see if game is associated with launcher before startup
Launchers, such as steam, need to boot up before the games that
use them can run.  For example, if Doom was purchased via Steam,
starting the process for Doom will invoke the Steam launcher
to start if it is not already open.  The launcher needs to be
running first before the game can run, but FireDoor was not
configured to wait on this process before reading temps.  This
means that FireDoor would have checked to see if the process for
the test app was running while the launcher was booting up, and
would abort because it has not found the process.  This issue has
been somewhat addressed by asking the user to check that their
launcher is running BEFORE running the game.  Future versions
will force the user to open the launcher before running games.