Pass a ProcessStartInfo object into CpuTempService
The program would return a ProcessStart object from TestAppService
and later pass it to the CpuTempService.  To help simplify things,
we are now returning a ProcessStartInfo from the TestAppService
and passing it into the CpuTempService.  The constructor in
the CpuTempService will then create the ProcessStart object
using the ProcessStartInfo object data.