Fix bug where incorrect max temp is displayed
Users are allowed to pick their own max tempratures.  However,
when their max temprature was exceeded, the program would inform
the user that the value of that temp was 60 degrees (despite it being
the temp they chose).  This commit addresses the bug by showing the
correct temp in the message.