Add Fody and Costura.Fody to allow DLLs to be included in EXE
Fody and Costura.Fody allow DLLs specified in the FodyWeavers.xml
file to be bundled into the FireDoor.exe file that gets generated
at each build.  This helps ensure that the OpenHardwareMonitorLib
dependency does not impact FireDoor when shipped as an EXE, as
Fody and Costura.Fody will have included it as part of the EXE.

A change to the location of the coreTempCSV file was made so that
it generates in C:\Temp.