Fix bug where test app's process was not killed
A bug that did not allow the test app to be closed was discovered.
This would happen when a CPU core would exceed the value of the max
temp immediately after the test app would start up.
This code change ensures that the app will be closed regardless
how close the kill command is sent after the app's start.