Add support for hours
FireDoor was only capable using minutes as its largest unit of
measurement for time.  When a program's uptime would surpass an
hour, the minutes would start back at 0.  For example, if an
program were to run for 1 hour and 20 minutes, FireDoor
would show that the program only ran for 20 minutes.  This
commit addresses that issue by allowing hours to be displayed.